Working on magnetometer calibration

+++ MCU_H7/cdp_H7_implementation/Core/Inc/magneto.h
@@ -93,7 +93,7 @@ typedef struct
   void      (*ClearIT)(void);
   void      (*FilterConfig)(uint8_t);
   void      (*FilterCmd)(uint8_t);
-  void      (*GetXYZ)(int16_t *);
+  void      (*GetXYZ)(float *);
 }MAGNETO_DrvTypeDef;
 /**
   * @}
